requirements.txt
# django
Django>=2.0

# django rest framework
djangorestframework>=3.9.0

#http://django-debug-toolbar.readthedocs.io/en/stable/
django-debug-toolbar>=1.9.1